+++ packages/openapi-generator/src/openapi-types.ts
@@ -24,6 +24,7 @@ export interface OpenApiOperation extends OpenAPIV3.OperationObject {
   path: string;
   requestBody?: OpenApiRequestBody;
   parameters: OpenApiParameter[];
+  'x-sap-cloud-sdk-operation-name'?: string;
 }
 
 const supportedMethods = {